refactor: float -> double

+++ src/main/java/com/threestar/trainus/domain/ranking/dto/RankingData.java
@@ -7,7 +7,7 @@ public interface RankingData {
 
 	Integer getReviewCount();
 
-	Float getRating();
+	Double getRating();
 
 	String getProfileImage();
 }
